Add comments for test helper functions. Add tests for the non rest case.
